fix: convert all fontSize values to double to prevent runtime type 'int' is not a subtype of 'double?' error

+++ lib/core/widgets/app_avatar.dart
@@ -92,7 +92,7 @@ class AppAvatar extends StatelessWidget {
         cleanInitials,
         style: typography.labelPill.copyWith(
           color: Colors.white,
-          fontSize: radius * 0.8,
+          fontSize: (radius * 0.8).toDouble(),
           fontWeight: FontWeight.bold,
         ),
       ),
